# Access control: Rewrite access control SQL filter