--- a/pkg/services/accesscontrol/filter_bench_test.go
+++ b/pkg/services/accesscontrol/filter_bench_test.go
@@ -13,9 +13,11 @@ import (
 	"github.com/grafana/grafana/pkg/services/sqlstore"
 )
 
-func BenchmarkFilter10_10(b *testing.B)   { benchmarkFilter(b, 10, 10) }
-func BenchmarkFilter100_10(b *testing.B)  { benchmarkFilter(b, 100, 10) }
-func BenchmarkFilter100_100(b *testing.B) { benchmarkFilter(b, 100, 100) }
+func BenchmarkFilter10_10(b *testing.B)     { benchmarkFilter(b, 10, 10) }
+func BenchmarkFilter100_10(b *testing.B)    { benchmarkFilter(b, 100, 10) }
+func BenchmarkFilter100_100(b *testing.B)   { benchmarkFilter(b, 100, 100) }
+func BenchmarkFilter1000_100(b *testing.B)  { benchmarkFilter(b, 1000, 100) }
+func BenchmarkFilter1000_1000(b *testing.B) { benchmarkFilter(b, 1000, 100) }
 
 func benchmarkFilter(b *testing.B, numDs, numPermissions int) {
 	store, permissions := setupFilterBenchmark(b, numDs, numPermissions)