--- a/pkg/services/accesscontrol/filter_test.go
+++ b/pkg/services/accesscontrol/filter_test.go
@@ -13,145 +13,6 @@ import (
 	"github.com/grafana/grafana/pkg/services/sqlstore"
 )
 
-type filterTest struct {
-	desc          string
-	driverName    string
-	sqlID         string
-	action        string
-	prefix        string
-	permissions   []*accesscontrol.Permission
-	expectedQuery string
-	expectedArgs  []interface{}
-}
-
-func TestFilter(t *testing.T) {
-	tests := []filterTest{
-		{
-			desc:       "should produce datasource filter with sqlite driver",
-			driverName: "sqlite3",
-			sqlID:      "data_source.id",
-			prefix:     "datasources",
-			action:     "datasources:query",
-			permissions: []*accesscontrol.Permission{
-				{Action: "datasources:query", Scope: "datasources:id:1"},
-				{Action: "datasources:query", Scope: "datasources:id:2"},
-				{Action: "datasources:query", Scope: "datasources:id:3"},
-				{Action: "datasources:query", Scope: "datasources:id:8"},
-				// Other permissions
-				{Action: "datasources:write", Scope: "datasources:id:100"},
-				{Action: "datasources:delete", Scope: "datasources:id:101"},
-			},
-			expectedQuery: `
-		? || ':id:' || data_source.id IN (
-			WITH t(scope) AS (
-				VALUES (?), (?), (?), (?)
-			)
-			SELECT IIF(t.scope = '*' OR t.scope = ? || ':*' OR t.scope = ? || ':id:*', ? || ':id:' || data_source.id, t.scope) FROM t
-		)
-	`,
-			expectedArgs: []interface{}{
-				"datasources",
-				"datasources:id:1",
-				"datasources:id:2",
-				"datasources:id:3",
-				"datasources:id:8",
-				"datasources",
-				"datasources",
-				"datasources",
-			},
-		},
-		{
-			desc:       "should produce dashboard filter with mysql driver",
-			driverName: "mysql",
-			sqlID:      "dashboard.id",
-			prefix:     "dashboards",
-			action:     "dashboards:read",
-			permissions: []*accesscontrol.Permission{
-				{Action: "dashboards:read", Scope: "dashboards:id:1"},
-				{Action: "dashboards:read", Scope: "dashboards:id:2"},
-				{Action: "dashboards:read", Scope: "dashboards:id:5"},
-				// Other permissions
-				{Action: "dashboards:write", Scope: "dashboards:id:100"},
-				{Action: "dashboards:delete", Scope: "dashboards:id:101"},
-			},
-			expectedQuery: `
-		CONCAT(?, ':id:', dashboard.id) IN (
-			SELECT IF(t.scope = '*' OR t.scope = CONCAT(?, ':*') OR t.scope = CONCAT(?, ':id:*'), CONCAT(?, ':id:', dashboard.id), t.scope) FROM
-			(SELECT ? AS scope UNION ALL SELECT ? UNION ALL SELECT ?) AS t
-		)
-	`,
-			expectedArgs: []interface{}{
-				"dashboards",
-				"dashboards",
-				"dashboards",
-				"dashboards",
-				"dashboards:id:1",
-				"dashboards:id:2",
-				"dashboards:id:5",
-			},
-		},
-		{
-			desc:       "should produce user filter with postgres driver",
-			driverName: "postgres",
-			sqlID:      "user.id",
-			prefix:     "users",
-			action:     "users:read",
-			permissions: []*accesscontrol.Permission{
-				{Action: "users:read", Scope: "users:id:1"},
-				{Action: "users:read", Scope: "users:id:100"},
-				// Other permissions
-				{Action: "dashboards:write", Scope: "dashboards:id:100"},
-				{Action: "dashboards:delete", Scope: "dashboards:id:101"},
-			},
-			expectedQuery: `
-		CONCAT(?, ':id:', user.id) IN (
-			SELECT
-				CASE WHEN p.scope = '*' OR p.scope = CONCAT(?, ':*') OR p.scope = CONCAT(?, ':id:*') THEN CONCAT(?, ':id:', user.id)
-				ELSE p.scope
-	    		END
-			FROM (VALUES (?), (?)) as p(scope)
-		)
-	`,
-			expectedArgs: []interface{}{
-				"users",
-				"users",
-				"users",
-				"users",
-				"users:id:1",
-				"users:id:100",
-			},
-		},
-	}
-
-	// set sqlIDAcceptList before running tests
-	restore := accesscontrol.SetAcceptListForTest(map[string]struct{}{
-		"user.id":        {},
-		"dashboard.id":   {},
-		"data_source.id": {},
-	})
-	defer restore()
-
-	for _, tt := range tests {
-		t.Run(tt.desc, func(t *testing.T) {
-			query, args, err := accesscontrol.Filter(
-				context.Background(),
-				FakeDriver{name: tt.driverName},
-				tt.sqlID,
-				tt.prefix,
-				tt.action,
-				&models.SignedInUser{OrgId: 1, Permissions: map[int64]map[string][]string{1: accesscontrol.GroupScopesByAction(tt.permissions)}},
-			)
-			require.NoError(t, err)
-			assert.Equal(t, tt.expectedQuery, query)
-
-			require.Len(t, args, len(tt.expectedArgs))
-			for i := range tt.expectedArgs {
-				assert.Equal(t, tt.expectedArgs[i], args[i])
-			}
-		})
-	}
-}
-
 type filterDatasourcesTestCase struct {
 	desc                string
 	sqlID               string
@@ -221,7 +82,6 @@ func TestFilter_Datasources(t *testing.T) {
 			baseSql := `SELECT data_source.* FROM data_source WHERE`
 			query, args, err := accesscontrol.Filter(
 				context.Background(),
-				&FakeDriver{name: "sqlite3"},
 				tt.sqlID,
 				"datasources",
 				"datasources:read",
@@ -243,12 +103,4 @@ func TestFilter_Datasources(t *testing.T) {
 			}
 		})
 	}
-}
-
-type FakeDriver struct {
-	name string
-}
-
-func (f FakeDriver) DriverName() string {
-	return f.name
 }