**What this PR does / why we need it**:
When working on dashboard and folder permissions i needed to change the search filter to use the access control SQL filter.
I noticed that it was pretty slow with lots of permissions so i rewrote the filter to make it more efficient.

Included in benchmarks is creating the SQL filter and fetching data sources based on that

Benchmark before rewrite:
```
benchmark                                    iter           time/iter      bytes alloc              allocs
---------                                    ----           ---------      -----------              ------
BenchmarkFilter10_10-8                       2733        467.37 μs/op       80667 B/op      2311 allocs/op
BenchmarkFilter100_10-8                      1822        623.34 μs/op       80661 B/op      2311 allocs/op
BenchmarkFilter100_100-8                      208       6829.00 μs/op      732045 B/op     21357 allocs/op
BenchmarkFilter1000_100-8                      43      26028.64 μs/op      731832 B/op     21356 allocs/op
BenchmarkFilter1000_1000-8                     43      26548.67 μs/op      732335 B/op     21359 allocs/op
```
Benchmark after rewrite:
```
benchmark                                    iter           time/iter      bytes alloc              allocs
---------                                    ----           ---------      -----------              ------
BenchmarkFilter10_10-8                       3306        379.39 μs/op       78878 B/op      2277 allocs/op
BenchmarkFilter100_10-8                      3442        370.94 μs/op       78862 B/op      2277 allocs/op
BenchmarkFilter100_100-8                      318       3251.47 μs/op      722990 B/op     21143 allocs/op
BenchmarkFilter1000_100-8                     355       3283.74 μs/op      723152 B/op     21144 allocs/op
BenchmarkFilter1000_1000-8                    315       3384.78 μs/op      723120 B/op     21144 allocs/op
```